@@ -1,35 +1,26 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import * as serviceWorker from './serviceWorker';
-import firebase from 'firebase/app';
+import { createRoot } from 'react-dom/client';
+import {
+  persistentLocalCache,
+  persistentMultipleTabManager
+} from 'firebase/firestore';
+import { initializeFirestore } from 'firebase/firestore';
 import 'firebase/auth';
-import 'firebase/firebase-messaging';
-import 'firebase/firestore';
+import 'firebase/messaging';
+import App, { app } from './App';
+import * as serviceWorker from './serviceWorker';
+import './index.css';
 
-  // Initialize Firebase services
-  firebase.initializeApp({
-    apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
-    authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
-    databaseURL: process.env.REACT_APP_FIREBASE_DATABASE_URL,
-    projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
-    storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
-    messagingSenderId: process.env.REACT_APP_FIREBASE_MESSAGING_SENDER_ID,
-    appId: process.env.REACT_APP_FIREBASE_APP_ID,
-  });
+// Enable Firestore database persistence so database can be queried/used offline
+export const firestore = initializeFirestore(app, {
+  localCache: persistentLocalCache(
+    /*settings*/ { tabManager: persistentMultipleTabManager() }
+  )
+});
 
-  // Enable Firestore database persistence so database can be queried/used offline
-  firebase.firestore().enablePersistence()
-  .catch(function(err) {
-      if (err.code === 'failed-precondition') {
-          console.log('Failed Firestore persistence pre-condition: ', err);
-      } else if (err.code === 'unimplemented') {
-          console.log('Browser doesn\'t support features needed for Firestore persistence.');
-      }
-  });
-
-ReactDOM.render(<App />, document.getElementById('root'));
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(<App />);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
